feat: v2 refactor and new features
- feat: add `uuid-kotlinx` module for kotlinx serialization
  integration library

- feat: implement kotlinx serialization for uuids

- chore: initial refactor into multi-module project

- chore: update lockfiles and verification metadata
