+++ gradle/libs.versions.toml
@@ -12,6 +12,7 @@ versionCheck = "0.51.0"
 versionCatalogUpdate = "0.8.4"
 testlogger = "4.0.0"
 nexus = "1.3.0"
+kotlinx-serialization = "1.6.2"
 sigstore = "0.5.0"
 spdx = "0.6.0"
 cyclonedx = "1.8.2"
@@ -39,5 +40,8 @@ buildTimeTracker = { id = "com.asarkar.gradle.build-time-tracker", version.ref =
 
 [libraries]
 ktlint = { group = "com.pinterest", name = "ktlint", version.ref = "ktlint" }
+kotlinx-serialization-core = { group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-core", version.ref = "kotlinx-serialization" }
+kotlinx-serialization-json = { group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-json", version.ref = "kotlinx-serialization" }
+kotlinx-serialization-protobuf = { group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-protobuf", version.ref = "kotlinx-serialization" }
 
 [bundles]
