+++ gradle/verification-metadata.xml
@@ -12,6 +12,7 @@
          <trust file=".*kotlin.*pom" regex="true"/>
          <trust file="kotlin-.*-prebuilt-.*[.]tar[.]gz" regex="true"/>
          <trust file="node-.*[.]tar[.]gz" regex="true"/>
+         <trust file="gradle-.*.zip" regex="true"/>
       </trusted-artifacts>
       <ignored-keys>
          <ignored-key id="4BE41A4907BCD034" reason="Key couldn't be downloaded from any key server"/>
