+++ settings.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Elide Ventures, LLC.
+ * Copyright (c) 2023-2024 Elide Technologies, Inc.
  *
  * Licensed under the MIT license (the "License"); you may not use this file except in compliance
  * with the License. You may obtain a copy of the License at
@@ -43,6 +43,10 @@ gradleEnterprise {
     termsOfServiceAgree = "yes"
   }
 }
+
+// KotlinX extensions and support package.
+include(":subprojects:uuid-core")
+include(":subprojects:uuid-kotlinx")
 
 enableFeaturePreview("STABLE_CONFIGURATION_CACHE")
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
