--- src/commonMain/kotlin/deprecations.kt
@@ -1,24 +1,0 @@
-/*
- * Copyright (c) 2023 Elide Ventures, LLC.
- *
- * Licensed under the MIT license (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
- *
- *   https://opensource.org/license/mit/
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under the License.
- */
-
-package dev.elide.uuid
-
-@Deprecated("Use `Uuid` instead.", ReplaceWith("Uuid"))
-public typealias UUID = Uuid
-
-@Deprecated(
-    message = "Use uuidFrom() instead. This will be removed in the next release.",
-    replaceWith = ReplaceWith("Uuid.bytes"),
-)
-public val Uuid.uuid: ByteArray
-    get() = bytes
