+++ subprojects/uuid-core/src/commonTest/data/uuid3.txt
