+++ subprojects/uuid-core/src/commonTest/data/uuid5.txt
