+++ subprojects/uuid-core/src/jvmMain/kotlin/namebased.kt
@@ -1,0 +1,59 @@
+/*
+ * Copyright (c) 2023-2024 Elide Technologies, Inc.
+ *
+ * Licensed under the MIT license (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ *   https://opensource.org/license/mit/
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under the License.
+ */
+
+package dev.elide.uuid
+
+import java.security.MessageDigest
+
+/**
+ * Constructs a "Name-Based" version 3 [UUID][Uuid].
+ *
+ * Version 3 UUIDs are created by combining a name and
+ * a namespace using the MD5 hash function.
+ *
+ * @param namespace for the "Name-Based" UUID
+ * @param name withing the namespace for the "Name-Based" UUID
+ * @return New version 3 [UUID][Uuid].
+ * @see <a href="https://tools.ietf.org/html/rfc4122#section-4.3">RFC 4122: Section 4.3</a>
+ */
+public fun uuid3Of(namespace: Uuid, name: String): Uuid =
+    nameBasedUuidOf(namespace, name, JvmHasher("MD5", 3))
+
+/**
+ * Constructs a "Name-Based" version 5 [UUID][Uuid].
+ *
+ * Version 5 UUIDs are created by combining a name and
+ * a namespace using the SHA-1 hash function.
+ *
+ * @param namespace for the "Name-Based" UUID
+ * @param name withing the namespace for the "Name-Based" UUID
+ * @return New version 5 [UUID][Uuid].
+ * @see <a href="https://tools.ietf.org/html/rfc4122#section-4.3">RFC 4122: Section 4.3</a>
+ */
+public fun uuid5Of(namespace: Uuid, name: String): Uuid =
+    nameBasedUuidOf(namespace, name, JvmHasher("SHA-1", 5))
+
+private class JvmHasher(
+    algorithmName: String,
+    override val version: Int,
+) : UuidHasher {
+    private val digest = MessageDigest.getInstance(algorithmName)
+
+    override fun update(input: ByteArray) {
+        digest.update(input)
+    }
+
+    override fun digest(): ByteArray {
+        return digest.digest()
+    }
+}
