+++ subprojects/uuid-core/src/jvmMain/kotlin/uuid.kt
@@ -1,0 +1,157 @@
+/*
+ * Copyright (c) 2023-2024 Elide Technologies, Inc.
+ *
+ * Licensed under the MIT license (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ *   https://opensource.org/license/mit/
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under the License.
+ */
+
+package dev.elide.uuid
+
+import java.nio.ByteBuffer
+import java.io.Serializable as JSerializable
+import java.util.UUID as JavaUUID
+
+/**
+ * Typealias for a UUID [Long] pair.
+ */
+public typealias UuidPair = Pair<Long, Long>
+
+/**
+ * A RFC4122 UUID wrapper class.
+ *
+ * This type wraps the standard [JavaUUID] type, and behaves identically; it will produce a string via [toString] in the
+ * same manner, and is [Comparable] with other UUIDs, and [java.io.Serializable].
+ */
+@JvmInline
+public actual value class Uuid public actual constructor (public val pair: Pair<Long, Long>):
+  Comparable<Uuid>,
+  JSerializable {
+  public companion object {
+    @Suppress("ConstPropertyName")
+    public const val serialVersionUID: Long = 1L
+
+    @JvmStatic public fun of(most: Long, least: Long): Uuid = Uuid(most to least)
+
+    @JvmStatic public fun from(javaUuid: JavaUUID): Uuid = Uuid(
+      javaUuid.mostSignificantBits to javaUuid.leastSignificantBits
+    )
+
+    @JvmStatic public fun randomUUID(): Uuid = from(JavaUUID.randomUUID())
+  }
+
+  public val javaUuid: JavaUUID get() = JavaUUID(mostSignificantBits, leastSignificantBits)
+
+  public actual constructor(msb: Long, lsb: Long): this(msb to lsb)
+
+  actual override fun compareTo(other: Uuid): Int =
+    javaUuid.compareTo(other.javaUuid)
+
+  @Suppress("NOTHING_TO_INLINE")
+  public actual inline val mostSignificantBits: Long get() = pair.first
+
+  @Suppress("NOTHING_TO_INLINE")
+  public actual inline val leastSignificantBits: Long get() = pair.second
+
+  override fun toString(): String = javaUuid.toString()
+}
+
+public actual inline val Uuid.version: Int
+  get() = javaUuid.version()
+
+public actual inline val Uuid.variant: Int
+  get() = javaUuid.variant()
+
+public actual val Uuid.bytes: ByteArray
+    get() = ByteBuffer.allocate(16).putLong(mostSignificantBits).putLong(leastSignificantBits).array()
+
+@Suppress("NOTHING_TO_INLINE")
+public actual inline fun uuid4(): Uuid = Uuid.randomUUID()
+
+public actual fun uuidOf(bytes: ByteArray): Uuid {
+  require(bytes.size == UUID_BYTES) {
+    "Invalid UUID bytes. Expected $UUID_BYTES bytes; found ${bytes.size}"
+  }
+  return ByteBuffer.wrap(bytes).let { Uuid.of(it.long, it.long) }
+}
+
+// Implementation Notes:
+// - This algorithm could be implemented in the common module and reused for all
+//   targets. However, using two longs is not necessarily the best performing
+//   solution in other targets. JavaScript for instance does not have native
+//   support for 64 bit integers and they are always wrapped in an object
+//   (similar to a BigDecimal in Java), a org.khronos.webgl.Uint8ClampedArray is
+//   probably the best choice there. Native targets are probably better at
+//   dealing with byte arrays, or even directly using native functions… only
+//   careful benchmarking can tell and UUIDs are ALWAYS in the critical path of
+//   applications, hence, we want to have the best solution for each target.
+// - We are not using any constants but rather literal values because constants
+//   lead to code being generated that is accessible via reflection, but we do
+//   not want to leak anything.
+// - The length verification at the beginning does not always lead to the
+//   exception being thrown because length is UTF-16 based and not byte based as
+//   it is actually required here. Careful benchmarking, however, showed that
+//   any kind of conversion leads to a loss in throughput that is not worth it
+//   to pay since the hex-conversion that follows is going to catch all chars
+//   that are not valid hexadecimal digits anyways.
+// - Regarding benchmarks, the following numbers are from a Zulu OpenJDK 13 run
+//   between java.util.UUID and this implementation on a 2017 MacBook Pro with
+//   Catalina:
+//
+//   Benchmark                    Mode  Cnt         Score        Error  Units
+//   UuidParserBenchmark.java    thrpt   25   9628223.641 ± 248719.696  ops/s
+//   UuidParserBenchmark.kotlin  thrpt   25  13948372.928 ± 694944.523  ops/s
+//
+//   Other strategies (e.g. ByteArray, ByteBuffer, …) are all slower than the
+//   MSB/LSB approach. The reason for that has probably to do with how the JVM
+//   works and that such arrays always lead to heap allocations.
+public actual fun uuidFrom(string: String): Uuid =
+  if (string.length == 36) {
+    Uuid.of(string.segmentToLong(0, 19), string.segmentToLong(19, 36))
+  } else {
+    throw IllegalArgumentException(
+      "Invalid UUID string, expected exactly 36 characters but got ${string.length}: $string",
+    )
+  }
+
+private fun String.segmentToLong(start: Int, end: Int): Long {
+  var result = 0L
+  var i = start
+  do {
+    if (this[i] == '-') {
+      require(i == 8 || i == 13 || i == 18 || i == 23) {
+        "Invalid UUID string, encountered dash at index $i but it can occur only at 8, 13, 18, or 23: $this"
+      }
+    } else {
+      result *= 16
+      when (this[i]) {
+        '0' -> Unit
+        '1' -> result += 1L
+        '2' -> result += 2L
+        '3' -> result += 3L
+        '4' -> result += 4L
+        '5' -> result += 5L
+        '6' -> result += 6L
+        '7' -> result += 7L
+        '8' -> result += 8L
+        '9' -> result += 9L
+        'a', 'A' -> result += 10L
+        'b', 'B' -> result += 11L
+        'c', 'C' -> result += 12L
+        'd', 'D' -> result += 13L
+        'e', 'E' -> result += 14L
+        'f', 'F' -> result += 15L
+        else -> throw IllegalArgumentException(
+          "Invalid UUID string, encountered non-hexadecimal digit `${this[i]}` at index $i in: $this",
+        )
+      }
+    }
+  } while (++i < end)
+
+  return result
+}
