+++ subprojects/uuid-core/src/mingwMain/kotlin/platform.kt
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2023-2024 Elide Technologies, Inc.
+ *
+ * Licensed under the MIT license (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ *   https://opensource.org/license/mit/
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under the License.
+ */
+
+package dev.elide.uuid
+
+import kotlinx.cinterop.addressOf
+import kotlinx.cinterop.convert
+import kotlinx.cinterop.reinterpret
+import kotlinx.cinterop.usePinned
+import platform.windows.BCRYPT_USE_SYSTEM_PREFERRED_RNG
+import platform.windows.BCryptGenRandom
+
+internal actual fun getRandomUuidBytes(): ByteArray {
+    val bytes = ByteArray(UUID_BYTES)
+    bytes.usePinned {
+        BCryptGenRandom(null, it.addressOf(0).reinterpret(), UUID_BYTES.toUInt(), BCRYPT_USE_SYSTEM_PREFERRED_RNG.convert())
+    }
+    return bytes
+}
